+++ Changelog.md
@@ -7,6 +7,8 @@ frameworks.
 
 _latest update: Wed Jul 29 23:10:24 CEST 2015_
 
+* Adding a check for CVE-2014-9490: Denial of Service in raven-ruby gem (issue
+  #124)
 * Adding a check for CVE-2015-3448: Information leakage in rest-client gem
   (issue #125)